Pushes branches from a remote to another
It is simple to use `git push --tags --all` but usually there a bunch of
stale commits everywhere and this default to not push branches that have
code older than 1 month old.

It should work as intended and be easy to apply changes or new options.
I was tempted to name this `git-migrate` or `git-sync` but seemed too
far fetched for the little (for the amount of code here) this does.